Add language to MVP.md about failure modes
Adds a "Failure" subsection to MVP.md briefly describing that
allocations can fail, that implementations are allowed to fail
however they wish up to and including terminating the host, and
noting this could open DoS vectors.

Instruction notes add a link to this section to warn these
instructions may trap or otherwise fail.

Closes out https://github.com/WebAssembly/gc/issues/122